Get outbound auth working for sesssions.
This works! I've come to the conclusion that we can't have
outbound auth done automatically the way I'd like for it to
be done, but I can at least take steps to make the process
as easy as I can make it for people. That's what future commits
will strive to do.