Separate logic for getting tpselector from an endpoint out of ast_sip_create_dialog().
This is step one towards being able to write the ast_sip_send_request() function. There
are several bits of logic that are in the middle of large functions that need to be separated
out into their own functions so that they can be called by ast_sip_send_request(). The
other things that need factoring out are

1) The determination of the remote URI to send a request to (currently in ast_sip_session_create_outbound)
2) Potentially, sip_dialog_create_from() needs to be written in such a way as to not be dialog-specific.
However, it's not really that dialog-specific to begin with.

With all that separated out, creating an out-of-dialog request will be a cakewalk.