Bridge API enhancements: Implement threading model for bridge management thread.
High level overview of the bridging module objects.

Bridge-manager: This thread handles odd jobs for all bridges in the
system.  The manager has a service request queue that tells it to see what
needs doing to a bridge.

Bridge: The bridge object does not have a thread associated with it.  It
simply directs where a bridge channel thread should put frames written
into the bridge.  The bridge has an action queue for general bridge
servicing.  The only actions currently posted to the queue are deferred
destruction of the softmix bridge private structure.

Bridge-channel: The bridge channel component of a bridge has a thread
associated with it that services the reads and writes between the channel
and the bridge.  The thread also performs any actions to the channel such
as playing a file or running the connected line interception routines.

Bridge-technology: The bridge technology component of a bridge gives the
bridge its basic characteristics for how media passes through the bridge.
The current technologies implemented are holding (For parking calls),
simple two party, softmix (For multiparty like ConfBridge).

(closes issue ASTERISK-21052)
Reported by: Matt Jordan

Review: https://reviewboard.asterisk.org/r/2391/