Add some progress towards using P-Asserted-Identity for connected line.
This is a half-effort here because I wanted to go down the path of writing
an actual parser for the header that PJSIP could automatically use when it
encounters the header. Unfortunately, it is not an easy task since most of
the handy functions for parsing are hidden away as private functions in
the PJSIP parsing code.