Change the way that P-Asserted-Identity and Remote-Party-ID headers are parsed.
Before, they were being parsed as strings and then run through a URI parser. This
will break down if there are any header parameters present. Now we parse the header
as a fromto header so that we have easy access to the URI and parameters. This
will come in handy once we start looking at privacy values in Remote-Party-ID headers.