Commit work thus far for bridge events for app_stasis
Interest in events for a given bridge by a stasis application is
conveyed via interest in channels that are currently in that bridge.
I.e. if channel A is watched by stasis application Z and channel A
enters bridge B, stasis application Z will get bridge entrance and exit
events for channel A in bridge B as well as all intervening events that
occur on bridge B.  Current bridge events are on par with events
produced by AMI with the addition of a bridge snapshot event in case
further processing is required.

Testing has not yet commenced, but it compiles.