Add reload support.
This makes it so that a new container of unsolicited MWI will replace the old
one when a reload occurs. This way, any changes to endpoints' mailbox settings
can be reflected properly after a reload.

The next step is to do an audit of references since I'm almost certain I've
got things pretty messed up at this point. The intention is for the stasis
subscription to be the driving force behind the destruction of the MWI
subscriptions.