Fix a crash seen during solicited MWI testing.
The issue stemmed from the fact that the state change to
PJSIP_EVSUB_STATE_TERMINATED does not necessarily correspond with
the destruction of the underlying dialog. However, we use this
state change as our means of facilitating subscription destruction,
because it's the best way we have of doing so. Unfortunately, since
the dialog still exists, it means that the distributor will attempt
to use the serializer saved on the dialog, which unfortunately
has become invalid. Attempting to use the invalid serializer results
in a crash.

The fix put in place is to add a function that allows for an application
to remove a serializer from a dialog. This is useful for a case like this
where the serializer may disappear before the dialog has actually been
destroyed. This way, if any further requests/responses arrive on the
dialog, the distributor will not attempt to use an invalid serializer.
Instead, there will be no serializer used at all, which is fine since
our application that cares about serialization is no longer involved.

With this change, I realize that some explanatory comments are necessary
in some of the code, so I will add those and then put this up for
review.