Change from doing a bridge removal to changing the state of the bridge_channel.
I still see a crash when attempting to do a blind transfer. I think the issue
might be that the bridge is trying to write to a bridge_channel that has disappeared.
This may point to a refcounting issue in my code, or it could be that I'm attempting
to do something in a bridge action that is not accounted for in the other bridging
code.