Make local channel optimization also use the swap method.
Added the additional way to optimize out local channels in more
circumstances.  Swap the Local;2 channel in BridgeB with the only peer
channel (A) in BridgeA.  This optimization strategy is needed for the
parking and conference bridges to keep bridge channel information
associated with the same call.

A -- BridgeA -- Local;1-Local;2 -- BridgeB -- B

* Extracted the ast_bridge_channel_peer() routine from the simple_bridge
so local channel swap optimization can utilize it as well.

* Implemented ast_bridge_move() as a side benefit to the swap
optimization.

* Refactored ast_bridge_local_optimized_out() to simplify its structure
because the swap optimization needs to be checked before the merge
optimization.

(issue ASTERISK-21058)
Review: https://reviewboard.asterisk.org/r/2474/