Add naive attended transfer implementation.
This is missing a couple of necessary items in order to function
properly:

1) I need the ability for a channel that is removed from a bridge
to run an arbitrary callback. Right now, I've faked this by queuing
a bridge action on a bridged channel.

2) I need the "virtual" channel construct. Right now, I need a local
channel and I need to be able to do operations with both ends of it
(place the ;1 in a bridge and masquerade the ;2 into a channel). Currently
I'm not able to do this.