Create a subbranch of transfers for SIP transfers.
This is a continuation of transfer work. I'm trying to keep
my initial transfer branch in its current state since it is
still in review. This subbranch will be where a conversion of
transfers in chan_sip will occur. The focus is to make chan_sip
remove most of its code in favor of calling ast_bridge_transfer_blind
or ast_bridge_transfer_attended.