Add sound indexing and tweak JSON as necessary
It was discovered during the course of working on this that the JSON
model currently set for sounds needed some tweaking to reflect reality
a bit better, so now it will contain a list of language/format pairs to
describe which formats and languages are available on the system.

This includes indexing of all sounds in formats registered to asterisk
and sound descriptions in any languages where the descriptions are
present. Currently, only English descriptions will be provided in
responses to Stasis-HTTP queries.