Further changes to handle_invite_replaces.
Use ast_raw_answer() instead of sending a 200 OK in place.
This way, CDR/CEL/Stasis/Manager events can be sent as
expected. 

We no longer use ast_quiet_chan so it's been removed.
Remove the "ignoring in a stupid way" case. "ignore" means
that we've received a retransmission. If we get a retransmission
we shouldn't answer. We need to just ignore the request,
so that's what we do now.