Add method to determine allowability of unreal channel optimization between bridges.
The idea here is that, given two bridges, the question of whether unreal channel
optimization is allowed between the bridges would be allowed.

The attended transfer code uses this to determine how to act when there are two
bridges involved. If optimization is allowed, then the type of optimization that
would be used should be used directly. If optimization is not allowed, then an
unreal channel will be used to link the two bridges together to complete the
transfer.

The method to determine optimization allowability is public since it may be
useful for other purposes, such as determining whether to send stasis messages
about local channel optimizations.