Add features configuration diff to working copy.
This may be a bad idea since features configuration has
not yet been approved on reviewboard. However, I need the
changes here in order to continue with this task, so I'm
putting this in tentatively. Once the final version of
features configuration is merged to trunk, I'll work on
getting any potential conflicts sorted out here.