Add applicationmap documentation.
There are a couple of oddball things to deal with here.

1. Configuration documentation doesn't seem to have a good way
of indicating that the option name can be whatever is desired. For
now I'm using "name" as the configuration option and clarifying in
the description how applicationmap items are configured.

2. The syntax provided in the XML does not get printed to the console
when performing a "config show help features applicationmap name".